requirements.txt
annotated-types==0.7.0
asttokens==3.0.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
feedparser==6.0.11
fonttools==4.62.1
idna==3.11
importlib_metadata==9.0.0
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
PyYAML==6.0.2
pandas==3.0.1
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
pyzmq==27.1.0
requests==2.33.0
seaborn==0.13.2
six==1.17.0
soupsieve==2.8.3
squarify==0.4.4
stack-data==0.6.3
tenacity==9.1.4
tornado==6.5.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vnai==2.4.0
vnstock==3.5.0
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
vnstock_ezchart==0.0.3
wcwidth==0.6.0
wordcloud==1.9.6
zipp==3.23.0
